requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
djangorestframework_simplejwt==4.8.0
PyJWT==2.1.0
Pillow
django-model-utils==4.1.1
django-filter